+++ sources/@roots/sage/package.json
@@ -145,6 +145,7 @@
   "dependencies": {
     "@roots/blade-loader": "workspace:sources/@roots/blade-loader",
     "@roots/bud": "workspace:sources/@roots/bud",
+    "@roots/bud-client": "workspace:sources/@roots/bud-client",
     "@roots/bud-framework": "workspace:sources/@roots/bud-framework",
     "@roots/bud-preset-wordpress": "workspace:sources/@roots/bud-preset-wordpress",
     "@roots/bud-support": "workspace:sources/@roots/bud-support",
