+++ sources/@roots/sage/src/client/dom-ready.ts
@@ -1,20 +1,3 @@
-/**
- * Calls once document has loaded.
- *
- * @remarks
- * Callback function may be async or sync
- *
- * @param onReady - callback function
- * @returns void
- */
-interface domReady {
-  (onReady: () => unknown | (() => Promise<unknown>)): void
-}
-
-const domReady: domReady = onReady => {
-  window.requestAnimationFrame(async function check() {
-    document.body ? await onReady() : window.requestAnimationFrame(check)
-  })
-}
+import domReady from '@roots/bud-client/dom-ready'
 
 export default domReady
